Fix and extend compartmentalisation:
1) Fix case of fetching multiple URLs, which didn't previously work because the
sandbox would exit after fetching the first URL. Have now added an explicit
TERMINATE_SANDBOX message to tell the sandbox to exit.

2) Fix missing propagation of errno. fetch(...) checks the errno value if
stat(...) fails and so when stat(...) calls are proxied to the privileged
parent, the errno value needs to be propagatedback to the sandbox.

3) Add support for ephemeral sandboxes via the additional SANDBOX_EPHEMERAL
make option (persistent sandboxing is the default).

4) Make debug output more consistent. It is now of the form, e.g. [HOST-pid]
(filename:line_no) XXXXX (or equivalently [SANDBOX-pid] ... for debugging
output generated by the sandbox).